requirements.txt
streamlit
gspread
oauth2client
streamlit-authenticator
plotly
google-auth
pandas